dictator patch gui clean up
